+++ Gruntfile.js
@@ -6,7 +6,6 @@ var modPath = '../../server_mods/com.wondible.pa.puppetmaster/'
 var stream = 'stable'
 var media = require('./lib/path').media(stream)
 var build = 'ui/main/shared/js/build.js'
-var common = 'ui/main/shared/js/common.js'
 var live_game = 'ui/mods/puppetmaster/live_game.js'
 var live_game_unit_alert = 'ui/mods/puppetmaster/live_game_unit_alert.js'
 var live_game_devmode = 'ui/mods/puppetmaster/live_game_devmode.js'
@@ -17,14 +16,6 @@ module.exports = function(grunt) {
   // Project configuration.
   grunt.initConfig({
     copy: {
-      common: {
-        files: [
-          {
-            src: media + common,
-            dest: common,
-          },
-        ],
-      },
       back: {
         files: [
           {
@@ -59,8 +50,7 @@ module.exports = function(grunt) {
               'CHANGELOG.md',
               'com.wondible.pa.puppetmaster/**',
               'ui/**',
-              'pa/**',
-              'server-script/**'],
+              'pa/**'],
             dest: modPath,
           },
         ],
@@ -131,14 +121,8 @@ module.exports = function(grunt) {
     }
   })
 
-  grunt.registerTask('hackCommon', 'Add mod hook to common.js', function() {
-    var text = grunt.file.read(common)
-    var ext = grunt.file.read('common_extensions.js')
-    grunt.file.write(common, text + ext)
-  })
-
   // Default task(s).
-  grunt.registerTask('default', ['proc', 'copy:common', 'hackCommon', 'copy:mod']);
+  grunt.registerTask('default', ['proc', 'copy:mod']);
 
 };